test/system: speed up podman ps --external
(cherry picked from commit d25dadf80a30b2cb5ed02592d239cbb6fd64ad4f)

The buildah buil kill trick is bad as we have to sleep and wait to aboid
flakes which takes time. Instead it is possible to redo this build part
manually with buildah commands. It is not trival and harder to
understand but it safes 2-3s so I think it is worth it.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>