test/system: speed up podman ps --external
The buildah buil kill trick is bad as we have to sleep and wait to aboid
flakes which takes time. Instead it is possible to redo this build part
manually with buildah commands. It is not trival and harder to
understand but it safes 2-3s so I think it is worth it.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit 350dfabf66636f376f174ca879e0857069d1d672)
Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>