fix(wsl): Non-English system users cannot use wsl machine
When checking if WSL is already installed, podman relies on the English string returned by WSL.
However, in non-English systems, WSL will return content in the current system language (such as Chinese).
To avoid this situation, we cannot simply rely on the content returned by WSL and need to manually check if the kernel exists.

Signed-off-by: Kevin Cui <bh@bugs.cc>