test: include podman-stressor
podman-stressor - A tool based on Podman, cgroupv2, and systemd that creates
a cgroupv2 namespace with limited resources according to user input and starts
'N' containers to ensure Podman behaves correctly.

Additionally, it is possible to enable a stress feature that uses the stress-ng
tool to stress each container in parallel and verify if the applications and
services running in the container environment are stable and secure.

This tool also ensures that containers do not exceed their resource limits,
even in the presence of memory, CPU, or other system interferences.

Furthermore, it is an excellent way to test container scalability and
interference between containers using Podman, allowing you to trigger 100,
1k, 5k, or even 1 million containers.

Why bash?

We just call existing programs, no need to add extra layer of python,
golang or rust for calling program at this moment.

Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>