test/system: enable parallel cp tests
Cp tests overall seem to be one of the slowest things of all system
tests. A local run of this file takes over 4 minutes. As such doing this
in parallel will help us a lot and there is no technical reason why we
cannot do that. The only issue is the reuse of the same container names
for all tests.

Thus this commit fixes all names, yes it is pretty bad here so this is a
very big diff. Additionally I gave each test its own unique cX num to
make debugging leaks easier, with that I know exactly which tests
"leaked" compared to only random names.

Local test time (12 threads) is now 90s compared to the 4m before.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>