[skip-ci] Packit: split rhel and centos-stream jobs
This allows centos stream builds to run for outside contributors without
write access to the repos.

This commit also include centos-stream-9 build jobs so we can compare
both centos-stream-9 and rhel-9 builds if required. This will also be
useful when we want to run tests on both centos stream and rhel envs
using their respective builds.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>