Clearly indicate names w/ URLencoded duplicates
Previously it was easy for a maintainer to change a Cirrus-CI task name
w/o realizing it would (later) cause a downstream task or the `artifacts`
task to fail.  This is due to duplication of task names in URLencoded
format.  Worse, `artifacts` task failures are inconsequential in PRs,
and not all tasks are executed for every PR context (i.e. due to `skip`
or `only_if` exclusions).  So induced name change failures may not be
noticed until after a PR merges.

Improve the visibility of the situation by adding comments to all
consequential names.  That is, task names where there is a URLencoded
duplicate reference elsewhere.

Signed-off-by: Chris Evich <cevich@redhat.com>