pkg/machine/e2e: improve timeout handling
In case of timeouts actually log the command again and make sure to send
SIGABRT to the process as go will create a useful stack strace where we
can see where things are hanging. It also kill the process unlike the
default Eventually().Should(Exit()) call the leaves the process around.

The output will be captured by default in the log so we just see the
stack trace there.

And while at it lower the timeout to 180s, nothing should take that long
and if somethings hangs we at least see it earlier.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>